improve state management
