+++ shell/config.go
@@ -1,13 +1,15 @@
 package shell
 
+//Config is the config for the client.
 type Config struct {
 }
 
-type ShellClient struct {
+//Client is the client itself. Since we already have access to the shell no real provisioning needs to be done
+type Client struct {
 }
 
 // Client configures and returns a fully initialized ShellClient
 func (c *Config) Client() (interface{}, error) {
-	var client ShellClient
+	var client Client
 	return &client, nil
 }
