+++ test/scripts/delete.sh
@@ -1,3 +1,9 @@
+#!/bin/bash
+echo "deleting..."
+echo "writing some error" >&2
+
 IN=$(cat)
-echo $IN
+echo "stdin: ${IN}" #the old state
+
+#business logic
 rm -rf ex.json
